scraper/requirements.txt
aiohappyeyeballs==2.3.4
aiohttp==3.10.0
aiosignal==1.3.1
asyncio==3.4.3
attrs==24.1.0
frozenlist==1.4.1
gitdb==4.0.11
GitPython==3.1.43
idna==3.7
multidict==6.0.5
schedule==1.2.2
smmap==5.0.1
yarl==1.9.4